refactor: [KB-113] split into separate components AddTransactionModal screens
